Move modules into separate files
