+++ src/lib.rs
@@ -55,18 +55,23 @@
 #[macro_use]
 extern crate matches;
 
-mod tables;
-mod levels;
+mod char_data;
+mod char_data_tables;
+mod explicit;
+mod implicit;
+mod prepare;
+mod level;
 
-pub use tables::{BidiClass, bidi_class, UNICODE_VERSION};
-pub use levels::Level;
-
-use BidiClass::*;
+pub use char_data::{BidiClass, bidi_class, UNICODE_VERSION};
+pub use prepare::LevelRun;
+pub use level::Level;
 
 use std::borrow::Cow;
 use std::cmp::{max, min};
 use std::iter::repeat;
 use std::ops::Range;
+
+use BidiClass::*;
 
 /// Output of `process_text`
 ///
@@ -161,11 +166,6 @@ pub fn reorder_line<'a>(text: &'a str, line: Range<usize>, levels: &[Level]) -> 
     result.into()
 }
 
-/// A maximal substring of characters with the same embedding level.
-///
-/// Represented as a range of byte indices.
-pub type LevelRun = Range<usize>;
-
 /// Find the level runs within a line and return them in visual order.
 ///
 /// `line` is a range of bytes indices within `levels`.
@@ -206,7 +206,7 @@ pub fn visual_runs(line: Range<usize>, levels: &[Level]) -> Vec<LevelRun> {
     // http://www.unicode.org/reports/tr9/#L2
 
     // Stop at the lowest *odd* level.
-    min_level = min_level.get_lowest_odd_level_le();
+    min_level = min_level.get_lowest_rtl_level_ge();
 
     while max_level >= min_level {
         // Look for the start of a sequence of consecutive runs of max_level or higher.
@@ -344,577 +344,12 @@ fn assign_levels_to_removed_chars(para_level: Level, classes: &[BidiClass], leve
     }
 }
 
-/// 3.3.2 Explicit Levels and Directions
-///
-/// http://www.unicode.org/reports/tr9/#Explicit_Levels_and_Directions
-mod explicit {
-    use levels;
-    use super::{BidiClass, Level};
-    use super::BidiClass::*;
-
-    /// Compute explicit embedding levels for one paragraph of text (X1-X8).
-    ///
-    /// `classes[i]` must contain the BidiClass of the char at byte index `i`,
-    /// for each char in `text`.
-    pub fn compute(
-        text: &str,
-        para_level: Level,
-        initial_classes: &[BidiClass],
-        levels: &mut [Level],
-        classes: &mut [BidiClass],
-    ) {
-        assert!(text.len() == initial_classes.len());
-
-        // http://www.unicode.org/reports/tr9/#X1
-        let mut stack = DirectionalStatusStack::new();
-        stack.push(para_level, OverrideStatus::Neutral);
-
-        let mut overflow_isolate_count = 0u32;
-        let mut overflow_embedding_count = 0u32;
-        let mut valid_isolate_count = 0u32;
-
-        for (i, c) in text.char_indices() {
-            match initial_classes[i] {
-                // Rules X2-X5c
-                RLE | LRE | RLO | LRO | RLI | LRI | FSI => {
-                    let char_is_rtl = match initial_classes[i] {
-                        RLE | RLO | RLI => true,
-                        _ => false,
-                    };
-
-                    let last_level = stack.last().level;
-                    let new_level = if char_is_rtl {
-                        last_level.get_next_rtl_level()
-                    } else {
-                        last_level.get_next_ltr_level()
-                    };
-
-                    // X5a-X5c: Isolate initiators get the level of the last entry on the stack.
-                    let is_isolate = matches!(initial_classes[i], RLI | LRI | FSI);
-                    if is_isolate {
-                        levels[i] = last_level;
-                        match stack.last().status {
-                            OverrideStatus::RTL => classes[i] = R,
-                            OverrideStatus::LTR => classes[i] = L,
-                            _ => {}
-                        }
-                    }
-
-                    if new_level.valid() && overflow_isolate_count == 0 &&
-                       overflow_embedding_count == 0 {
-                        stack.push(
-                            new_level,
-                            match initial_classes[i] {
-                                RLO => OverrideStatus::RTL,
-                                LRO => OverrideStatus::LTR,
-                                RLI | LRI | FSI => OverrideStatus::Isolate,
-                                _ => OverrideStatus::Neutral,
-                            },
-                        );
-                        if is_isolate {
-                            valid_isolate_count += 1;
-                        } else {
-                            // The spec doesn't explicitly mention this step, but it is necessary.
-                            // See the reference implementations for comparison.
-                            levels[i] = new_level;
-                        }
-                    } else if is_isolate {
-                        overflow_isolate_count += 1;
-                    } else if overflow_isolate_count == 0 {
-                        overflow_embedding_count += 1;
-                    }
-                }
-                // http://www.unicode.org/reports/tr9/#X6a
-                PDI => {
-                    if overflow_isolate_count > 0 {
-                        overflow_isolate_count -= 1;
-                    } else if valid_isolate_count > 0 {
-                        overflow_embedding_count = 0;
-                        loop {
-                            // Pop everything up to and including the last Isolate status.
-                            match stack.vec.pop() {
-                                Some(Status { status: OverrideStatus::Isolate, .. }) => break,
-                                None => break,
-                                _ => continue,
-                            }
-                        }
-                        valid_isolate_count -= 1;
-                    }
-                    let last = stack.last();
-                    levels[i] = last.level;
-                    match last.status {
-                        OverrideStatus::RTL => classes[i] = R,
-                        OverrideStatus::LTR => classes[i] = L,
-                        _ => {}
-                    }
-                }
-                // http://www.unicode.org/reports/tr9/#X7
-                PDF => {
-                    if overflow_isolate_count > 0 {
-                        continue;
-                    }
-                    if overflow_embedding_count > 0 {
-                        overflow_embedding_count -= 1;
-                        continue;
-                    }
-                    if stack.last().status != OverrideStatus::Isolate && stack.vec.len() >= 2 {
-                        stack.vec.pop();
-                    }
-                    // The spec doesn't explicitly mention this step, but it is necessary.
-                    // See the reference implementations for comparison.
-                    levels[i] = stack.last().level;
-                }
-                // http://www.unicode.org/reports/tr9/#X6
-                B | BN => {}
-                _ => {
-                    let last = stack.last();
-                    levels[i] = last.level;
-                    match last.status {
-                        OverrideStatus::RTL => classes[i] = R,
-                        OverrideStatus::LTR => classes[i] = L,
-                        _ => {}
-                    }
-                }
-            }
-            // Handle multi-byte characters.
-            for j in 1..c.len_utf8() {
-                levels[i + j] = levels[i];
-                classes[i + j] = classes[i];
-            }
-        }
-    }
-
-    /// Entries in the directional status stack:
-    struct Status {
-        level: Level,
-        status: OverrideStatus,
-    }
-
-    #[derive(PartialEq)]
-    enum OverrideStatus {
-        Neutral,
-        RTL,
-        LTR,
-        Isolate,
-    }
-
-    struct DirectionalStatusStack {
-        vec: Vec<Status>,
-    }
-
-    impl DirectionalStatusStack {
-        fn new() -> Self {
-            DirectionalStatusStack { vec: Vec::with_capacity(levels::MAX_DEPTH as usize + 2) }
-        }
-        fn push(&mut self, level: Level, status: OverrideStatus) {
-            self.vec
-                .push(
-                    Status {
-                        level: level,
-                        status: status,
-                    },
-                );
-        }
-        fn last(&self) -> &Status {
-            self.vec.last().unwrap()
-        }
-    }
-}
-
-/// 3.3.3 Preparations for Implicit Processing
-///
-/// http://www.unicode.org/reports/tr9/#Preparations_for_Implicit_Processing
-mod prepare {
-    use super::{BidiClass, Level, LevelRun, max};
-    use super::BidiClass::*;
-
-    /// Output of `isolating_run_sequences` (steps X9-X10)
-    pub struct IsolatingRunSequence {
-        pub runs: Vec<LevelRun>,
-        pub sos: BidiClass, // Start-of-sequence type.
-        pub eos: BidiClass, // End-of-sequence type.
-    }
-
-    /// Compute the set of isolating run sequences.
-    ///
-    /// An isolating run sequence is a maximal sequence of level runs such that for all level runs
-    /// except the last one in the sequence, the last character of the run is an isolate initiator
-    /// whose matching PDI is the first character of the next level run in the sequence.
-    ///
-    /// Note: This function does *not* return the sequences in order by their first characters.
-    pub fn isolating_run_sequences(
-        para_level: Level,
-        initial_classes: &[BidiClass],
-        levels: &[Level],
-    ) -> Vec<IsolatingRunSequence> {
-        let runs = level_runs(levels, initial_classes);
-
-        // Compute the set of isolating run sequences.
-        // http://www.unicode.org/reports/tr9/#BD13
-
-        let mut sequences = Vec::with_capacity(runs.len());
-
-        // When we encounter an isolate initiator, we push the current sequence onto the
-        // stack so we can resume it after the matching PDI.
-        let mut stack = vec![Vec::new()];
-
-        for run in runs {
-            assert!(run.len() > 0);
-            assert!(stack.len() > 0);
-
-            let start_class = initial_classes[run.start];
-            let end_class = initial_classes[run.end - 1];
-
-            let mut sequence = if start_class == PDI && stack.len() > 1 {
-                // Continue a previous sequence interrupted by an isolate.
-                stack.pop().unwrap()
-            } else {
-                // Start a new sequence.
-                Vec::new()
-            };
-
-            sequence.push(run);
-
-            if matches!(end_class, RLI | LRI | FSI) {
-                // Resume this sequence after the isolate.
-                stack.push(sequence);
-            } else {
-                // This sequence is finished.
-                sequences.push(sequence);
-            }
-        }
-        // Pop any remaning sequences off the stack.
-        sequences.extend(stack.into_iter().rev().filter(|seq| seq.len() > 0));
-
-        // Determine the `sos` and `eos` class for each sequence.
-        // http://www.unicode.org/reports/tr9/#X10
-        return sequences
-                   .into_iter()
-                   .map(
-            |sequence| {
-                assert!(!sequence.len() > 0);
-                let start = sequence[0].start;
-                let end = sequence[sequence.len() - 1].end;
-
-                // Get the level inside these level runs.
-                let level = levels[start];
-
-                // Get the level of the last non-removed char before the runs.
-                let pred_level = match initial_classes[..start]
-                          .iter()
-                          .rposition(not_removed_by_x9) {
-                    Some(idx) => levels[idx],
-                    None => para_level,
-                };
-
-                // Get the level of the next non-removed char after the runs.
-                let succ_level = if matches!(initial_classes[end - 1], RLI | LRI | FSI) {
-                    para_level
-                } else {
-                    match initial_classes[end..].iter().position(not_removed_by_x9) {
-                        Some(idx) => levels[idx],
-                        None => para_level,
-                    }
-                };
-
-                IsolatingRunSequence {
-                    runs: sequence,
-                    sos: max(level, pred_level).get_bidi_class(),
-                    eos: max(level, succ_level).get_bidi_class(),
-                }
-            },
-        )
-                   .collect();
-    }
-
-    /// Finds the level runs in a paragraph.
-    ///
-    /// http://www.unicode.org/reports/tr9/#BD7
-    fn level_runs(levels: &[Level], classes: &[BidiClass]) -> Vec<LevelRun> {
-        assert!(levels.len() == classes.len());
-
-        let mut runs = Vec::new();
-        if levels.len() == 0 {
-            return runs;
-        }
-
-        let mut current_run_level = levels[0];
-        let mut current_run_start = 0;
-
-        for i in 1..levels.len() {
-            if !removed_by_x9(classes[i]) {
-                if levels[i] != current_run_level {
-                    // End the last run and start a new one.
-                    runs.push(current_run_start..i);
-                    current_run_level = levels[i];
-                    current_run_start = i;
-                }
-            }
-        }
-        runs.push(current_run_start..levels.len());
-        runs
-    }
-
-    /// Should this character be ignored in steps after X9?
-    ///
-    /// http://www.unicode.org/reports/tr9/#X9
-    pub fn removed_by_x9(class: BidiClass) -> bool {
-        matches!(class, RLE | LRE | RLO | LRO | PDF | BN)
-    }
-
-    // For use as a predicate for `position` / `rposition`
-    pub fn not_removed_by_x9(class: &BidiClass) -> bool {
-        !removed_by_x9(*class)
-    }
-
-    #[cfg(test)]
-    #[test]
-    fn test_level_runs() {
-        use levels;
-        let levels = &[0, 0, 0, 1, 1, 2, 0, 0];
-        assert_eq!(
-            level_runs(&levels::gen_vec(levels), &[L; 8]),
-            &[0..3, 3..5, 5..6, 6..8]
-        );
-    }
-
-    #[cfg(test)]
-    #[cfg_attr(rustfmt, rustfmt_skip)]
-    #[test]
-    fn test_isolating_run_sequences() {
-        use levels;
-        // Example 3 from http://www.unicode.org/reports/tr9/#BD13:
-
-        //              0    1   2    3  4  5  6    7   8    9 10
-        let classes = &[L, RLI, AL, LRI, L, R, L, PDI, AL, PDI, L];
-        let levels =  &[0,   0,  1,   1, 2, 3, 2,   1,  1,   0, 0];
-        let para_level = Level(0);
-
-        let sequences = isolating_run_sequences(para_level, classes, &levels::gen_vec(levels));
-        let runs: Vec<Vec<LevelRun>> = sequences.iter().map(|s| s.runs.clone()).collect();
-        assert_eq!(
-            runs,
-            vec![vec![4..5], vec![5..6], vec![6..7], vec![2..4, 7..9], vec![0..2, 9..11]]
-        );
-    }
-}
-
-/// 3.3.4 - 3.3.6. Resolve implicit levels and types.
-mod implicit {
-    use super::{BidiClass, Level, LevelRun, max};
-    use super::BidiClass::*;
-    use super::prepare::{IsolatingRunSequence, not_removed_by_x9, removed_by_x9};
-
-    /// 3.3.4 Resolving Weak Types
-    ///
-    /// http://www.unicode.org/reports/tr9/#Resolving_Weak_Types
-    pub fn resolve_weak(sequence: &IsolatingRunSequence, classes: &mut [BidiClass]) {
-        // FIXME (#8): This function applies steps W1-W6 in a single pass.  This can produce
-        // incorrect results in cases where a "later" rule changes the value of `prev_class` seen
-        // by an "earlier" rule.  We should either split this into separate passes, or preserve
-        // extra state so each rule can see the correct previous class.
-
-        let mut prev_class = sequence.sos;
-        let mut last_strong_is_al = false;
-        let mut et_run_indices = Vec::new(); // for W5
-
-        // Like sequence.runs.iter().flat_map(Clone::clone), but make indices itself clonable.
-        fn id(x: LevelRun) -> LevelRun {
-            x
-        }
-        let mut indices = sequence
-            .runs
-            .iter()
-            .cloned()
-            .flat_map(id as fn(LevelRun) -> LevelRun);
-
-        while let Some(i) = indices.next() {
-            match classes[i] {
-                // http://www.unicode.org/reports/tr9/#W1
-                NSM => {
-                    classes[i] = match prev_class {
-                        RLI | LRI | FSI | PDI => ON,
-                        _ => prev_class,
-                    };
-                }
-                EN => {
-                    if last_strong_is_al {
-                        // W2. If previous strong char was AL, change EN to AN.
-                        classes[i] = AN;
-                    } else {
-                        // W5. If a run of ETs is adjacent to an EN, change the ETs to EN.
-                        for j in &et_run_indices {
-                            classes[*j] = EN;
-                        }
-                        et_run_indices.clear();
-                    }
-                }
-                // http://www.unicode.org/reports/tr9/#W3
-                AL => classes[i] = R,
-
-                // http://www.unicode.org/reports/tr9/#W4
-                ES | CS => {
-                    let next_class = indices
-                        .clone()
-                        .map(|j| classes[j])
-                        .filter(not_removed_by_x9)
-                        .next()
-                        .unwrap_or(sequence.eos);
-                    classes[i] = match (prev_class, classes[i], next_class) {
-                        (EN, ES, EN) | (EN, CS, EN) => EN,
-                        (AN, CS, AN) => AN,
-                        (_, _, _) => ON,
-                    }
-                }
-                // http://www.unicode.org/reports/tr9/#W5
-                ET => {
-                    match prev_class {
-                        EN => classes[i] = EN,
-                        _ => et_run_indices.push(i), // In case this is followed by an EN.
-                    }
-                }
-                class => {
-                    if removed_by_x9(class) {
-                        continue;
-                    }
-                }
-            }
-
-            prev_class = classes[i];
-            match prev_class {
-                L | R => {
-                    last_strong_is_al = false;
-                }
-                AL => {
-                    last_strong_is_al = true;
-                }
-                _ => {}
-            }
-            if prev_class != ET {
-                // W6. If we didn't find an adjacent EN, turn any ETs into ON instead.
-                for j in &et_run_indices {
-                    classes[*j] = ON;
-                }
-                et_run_indices.clear();
-            }
-        }
-
-        // W7. If the previous strong char was L, change EN to L.
-        let mut last_strong_is_l = sequence.sos == L;
-        for run in &sequence.runs {
-            for i in run.clone() {
-                match classes[i] {
-                    EN if last_strong_is_l => {
-                        classes[i] = L;
-                    }
-                    L => {
-                        last_strong_is_l = true;
-                    }
-                    R | AL => {
-                        last_strong_is_l = false;
-                    }
-                    _ => {}
-                }
-            }
-        }
-    }
-
-    /// 3.3.5 Resolving Neutral Types
-    ///
-    /// http://www.unicode.org/reports/tr9/#Resolving_Neutral_Types
-    pub fn resolve_neutral(
-        sequence: &IsolatingRunSequence,
-        levels: &[Level],
-        classes: &mut [BidiClass],
-    ) {
-        let mut indices = sequence.runs.iter().flat_map(Clone::clone);
-        let mut prev_class = sequence.sos;
-
-        // Neutral or Isolate formatting characters (NI).
-        // http://www.unicode.org/reports/tr9/#NI
-        fn ni(class: BidiClass) -> bool {
-            matches!(class, B | S | WS | ON | FSI | LRI | RLI | PDI)
-        }
-
-        while let Some(mut i) = indices.next() {
-            // N0. Process bracket pairs.
-            // TODO
-
-            // Process sequences of NI characters.
-            let mut ni_run = Vec::new();
-            if ni(classes[i]) {
-                // Consume a run of consecutive NI characters.
-                ni_run.push(i);
-                let mut next_class;
-                loop {
-                    match indices.next() {
-                        Some(j) => {
-                            i = j;
-                            if removed_by_x9(classes[i]) {
-                                continue;
-                            }
-                            next_class = classes[j];
-                            if ni(next_class) {
-                                ni_run.push(i);
-                            } else {
-                                break;
-                            }
-                        }
-                        None => {
-                            next_class = sequence.eos;
-                            break;
-                        }
-                    };
-                }
-
-                // N1-N2.
-                let new_class = match (prev_class, next_class) {
-                    (L, L) => L,
-                    (R, R) | (R, AN) | (R, EN) | (AN, R) | (AN, AN) | (AN, EN) | (EN, R) |
-                    (EN, AN) | (EN, EN) => R,
-                    (_, _) => levels[i].get_bidi_class(),
-                };
-                for j in &ni_run {
-                    classes[*j] = new_class;
-                }
-                ni_run.clear();
-            }
-            prev_class = classes[i];
-        }
-    }
-
-    /// 3.3.6 Resolving Implicit Levels
-    ///
-    /// Returns the maximum embedding level in the paragraph.
-    ///
-    /// http://www.unicode.org/reports/tr9/#Resolving_Implicit_Levels
-    pub fn resolve_levels(classes: &[BidiClass], levels: &mut [Level]) -> Level {
-        let mut max_level = Level(0);
-
-        assert!(classes.len() == levels.len());
-        for i in 0..levels.len() {
-            match (levels[i].is_rtl(), classes[i]) {
-                // http://www.unicode.org/reports/tr9/#I1
-                (false, R) => levels[i].raise(1),
-                (false, AN) | (false, EN) => levels[i].raise(2),
-                // http://www.unicode.org/reports/tr9/#I2
-                (true, L) | (true, EN) | (true, AN) => levels[i].raise(1),
-                (_, _) => {}
-            }
-            max_level = max(max_level, levels[i]);
-        }
-        max_level
-    }
-}
-
 #[cfg(test)]
 mod test {
-    use super::BidiClass::*;
+    use super::*;
 
     #[test]
     fn test_initial_scan() {
-        use super::{InitialProperties, Level, initial_scan, ParagraphInfo};
-
         assert_eq!(
             initial_scan("a1", None),
             InitialProperties {
@@ -976,23 +411,11 @@ mod test {
     }
 
     #[test]
-    fn test_bidi_class() {
-        use super::bidi_class;
-
-        assert_eq!(bidi_class('c'), L);
-        assert_eq!(bidi_class('\u{05D1}'), R);
-        assert_eq!(bidi_class('\u{0627}'), AL);
-    }
-
-    #[test]
     fn test_process_text() {
-        use levels;
-        use super::{BidiInfo, Level, ParagraphInfo, process_text};
-
         assert_eq!(
             process_text("abc123", Some(Level(0))),
             BidiInfo {
-                levels: levels::gen_vec(&[0, 0, 0, 0, 0, 0]),
+                levels: Level::gen_vec(&[0, 0, 0, 0, 0, 0]),
                 classes: vec![L, L, L, EN, EN, EN],
                 paragraphs: vec![
                     ParagraphInfo {
@@ -1005,7 +428,7 @@ mod test {
         assert_eq!(
             process_text("abc אבג", Some(Level(0))),
             BidiInfo {
-                levels: levels::gen_vec(&[0, 0, 0, 0, 1, 1, 1, 1, 1, 1]),
+                levels: Level::gen_vec(&[0, 0, 0, 0, 1, 1, 1, 1, 1, 1]),
                 classes: vec![L, L, L, WS, R, R, R, R, R, R],
                 paragraphs: vec![
                     ParagraphInfo {
@@ -1018,7 +441,7 @@ mod test {
         assert_eq!(
             process_text("abc אבג", Some(Level(1))),
             BidiInfo {
-                levels: levels::gen_vec(&[2, 2, 2, 1, 1, 1, 1, 1, 1, 1]),
+                levels: Level::gen_vec(&[2, 2, 2, 1, 1, 1, 1, 1, 1, 1]),
                 classes: vec![L, L, L, WS, R, R, R, R, R, R],
                 paragraphs: vec![
                     ParagraphInfo {
@@ -1031,7 +454,7 @@ mod test {
         assert_eq!(
             process_text("אבג abc", Some(Level(0))),
             BidiInfo {
-                levels: levels::gen_vec(&[1, 1, 1, 1, 1, 1, 0, 0, 0, 0]),
+                levels: Level::gen_vec(&[1, 1, 1, 1, 1, 1, 0, 0, 0, 0]),
                 classes: vec![R, R, R, R, R, R, WS, L, L, L],
                 paragraphs: vec![
                     ParagraphInfo {
@@ -1044,7 +467,7 @@ mod test {
         assert_eq!(
             process_text("אבג abc", None),
             BidiInfo {
-                levels: levels::gen_vec(&[1, 1, 1, 1, 1, 1, 1, 2, 2, 2]),
+                levels: Level::gen_vec(&[1, 1, 1, 1, 1, 1, 1, 2, 2, 2]),
                 classes: vec![R, R, R, R, R, R, WS, L, L, L],
                 paragraphs: vec![
                     ParagraphInfo {
@@ -1057,7 +480,7 @@ mod test {
         assert_eq!(
             process_text("غ2ظ א2ג", Some(Level(0))),
             BidiInfo {
-                levels: levels::gen_vec(&[1, 1, 2, 1, 1, 1, 1, 1, 2, 1, 1]),
+                levels: Level::gen_vec(&[1, 1, 2, 1, 1, 1, 1, 1, 2, 1, 1]),
                 classes: vec![AL, AL, EN, AL, AL, WS, R, R, EN, R, R],
                 paragraphs: vec![
                     ParagraphInfo {
@@ -1071,7 +494,7 @@ mod test {
             process_text("a א.\nג", None),
             BidiInfo {
                 classes: vec![L, WS, R, R, CS, B, R, R],
-                levels: levels::gen_vec(&[0, 0, 1, 1, 0, 0, 1, 1]),
+                levels: Level::gen_vec(&[0, 0, 1, 1, 0, 0, 1, 1]),
                 paragraphs: vec![
                     ParagraphInfo {
                         range: 0..6,
@@ -1088,8 +511,6 @@ mod test {
 
     #[test]
     fn test_reorder_line() {
-        use super::{process_text, reorder_line};
-        use std::borrow::Cow;
         fn reorder(s: &str) -> Cow<str> {
             let info = process_text(s, None);
             let para = &info.paragraphs[0];
@@ -1123,27 +544,5 @@ mod test {
             reorder("Hello, \u{2068}\u{202E}world\u{202C}\u{2069}!"),
             "Hello, \u{2068}\u{202E}\u{202C}dlrow\u{2069}!"
         );
-    }
-
-    #[test]
-    fn test_removed_by_x9() {
-        use prepare::removed_by_x9;
-        let rem_classes = &[RLE, LRE, RLO, LRO, PDF, BN];
-        let not_classes = &[L, RLI, AL, LRI, PDI];
-        for x in rem_classes {
-            assert_eq!(removed_by_x9(*x), true);
-        }
-        for x in not_classes {
-            assert_eq!(removed_by_x9(*x), false);
-        }
-    }
-
-    #[test]
-    fn test_not_removed_by_x9() {
-        use prepare::not_removed_by_x9;
-        let non_x9_classes = &[L, R, AL, EN, ES, ET, AN, CS, NSM, B, S, WS, ON, LRI, RLI, FSI, PDI];
-        for x in non_x9_classes {
-            assert_eq!(not_removed_by_x9(&x), true);
-        }
     }
 }
