+++ tools/generate.py
@@ -150,14 +150,12 @@ def emit_table(f, name, t_data, t_type = "&'static [(char, char)]", is_pub=True,
     f.write("\n];\n\n")
 
 def emit_bidi_module(f, bidi_class, cats):
-    f.write("""pub use self::BidiClass::*;
+    f.write("""use self::BidiClass::*;
 
 #[allow(non_camel_case_types)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 /// Represents the Unicode character property **Bidi_Class**, also known as
 /// the *bidirectional character type*.
-///
-/// Use the `bidi_class` function to look up the BidiClass of a code point.
 ///
 /// http://www.unicode.org/reports/tr9/#Bidirectional_Character_Types
 pub enum BidiClass {
@@ -166,37 +164,19 @@ pub enum BidiClass {
         f.write("    " + cat + ",\n")
     f.write("""}
 
-fn bsearch_range_value_table(c: char, r: &'static [(char, char, BidiClass)]) -> BidiClass {
-    use ::std::cmp::Ordering::{Equal, Less, Greater};
-    match r.binary_search_by(|&(lo, hi, _)| {
-        if lo <= c && c <= hi { Equal }
-        else if hi < c { Less }
-        else { Greater }
-    }) {
-        Ok(idx) => {
-            let (_, _, cat) = r[idx];
-            cat
-        }
-        // UCD/extracted/DerivedBidiClass.txt: "All code points not explicitly listed
-        // for Bidi_Class have the value Left_To_Right (L)."
-        Err(_) => L
-    }
-}
-
-/// Find the BidiClass of a single char.
-pub fn bidi_class(c: char) -> BidiClass {
-    bsearch_range_value_table(c, bidi_class_table)
-}
-
 """)
 
-    emit_table(f, "bidi_class_table", bidi_class, "&'static [(char, char, BidiClass)]",
+    emit_table(
+        f,
+        "bidi_class_table",
+        bidi_class,
+        "&'static [(char, char, BidiClass)]",
         pfun=lambda x: "(%s,%s,%s)" % (escape_char(x[0]), escape_char(x[1]), x[2]),
-        is_pub=False)
+    )
 
 if __name__ == "__main__":
     os.chdir("../src/") # changing download path to /unicode-bidi/src/
-    r = "tables.rs"
+    r = "char_data_tables.rs"
     # downloading the test case files
     # fetch("BidiTest.txt")
     # fetch("BidiCharacterTest.txt")
